error fixed

+++ searchengine/src/main/java/com/example/searchengine/Ranker/Repository/LinksFromIndexerrepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.searchengine.Ranker.Entities.LinksFromIndexer;
 
-public interface LinksFromIndexerrepository extends JpaRepository<LinksFromIndexer, Long> {
+public interface LinksFromIndexerrepository extends JpaRepository<LinksFromIndexer, Integer> {
     // This interface is a repository for the LinksFromIndexer entity.
     // It extends JpaRepository to provide CRUD operations.
     // You can add custom query methods here if needed.
